requirements.txt
asgiref==3.5.2
cfgv==3.3.1
distlib==0.3.5
Django==4.1
django-versatileimagefield==2.2
filelock==3.8.0
gunicorn==20.1.0
identify==2.5.3
mypy==0.812
nodeenv==1.7.0
Pillow==9.2.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
python-magic==0.4.27
PyYAML==6.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing_extensions==4.3.0
virtualenv==20.16.3
